refactor: remove deprecated onboarding types and logic
