+++ backend/internal/services/playwright_service.go
@@ -12,41 +12,15 @@ import (
 )
 
 type PlaywrightService struct {
-	settingsService *SettingsService
 	apiKeyService   *ApiKeyService
 	userService     *UserService
 }
 
-func NewPlaywrightService(settingsService *SettingsService, apiKeyService *ApiKeyService, userService *UserService) *PlaywrightService {
+func NewPlaywrightService(apiKeyService *ApiKeyService, userService *UserService) *PlaywrightService {
 	return &PlaywrightService{
-		settingsService: settingsService,
 		apiKeyService:   apiKeyService,
 		userService:     userService,
 	}
-}
-
-func (ps *PlaywrightService) SkipOnboarding(ctx context.Context) error {
-	slog.Info("Playwright: Skipping onboarding for test environment")
-
-	err := ps.settingsService.SetBoolSetting(ctx, "onboardingCompleted", true)
-	if err != nil {
-		return fmt.Errorf("failed to skip onboarding: %w", err)
-	}
-
-	slog.Info("Playwright: Onboarding skipped successfully")
-	return nil
-}
-
-func (ps *PlaywrightService) ResetOnboarding(ctx context.Context) error {
-	slog.Info("Playwright: Resetting onboarding for test environment")
-
-	err := ps.settingsService.SetBoolSetting(ctx, "onboardingCompleted", false)
-	if err != nil {
-		return fmt.Errorf("failed to skip onboarding: %w", err)
-	}
-
-	slog.Info("Playwright: Onboarding reset successfully")
-	return nil
 }
 
 func (ps *PlaywrightService) CreateTestApiKeys(ctx context.Context, count int) ([]*apikey.ApiKeyCreatedDto, error) {
