+++ tests/setup/auth.setup.ts
@@ -1,12 +1,9 @@
 import { test as setup } from '@playwright/test';
-import { skipOnboarding } from '../utils/onboarding.util';
 import authUtil from '../utils/auth.util';
 
 const authFile = '.auth/login.json';
 
 setup('authenticate', async ({ page }) => {
-  await skipOnboarding(); // Deprecated
-
   await authUtil.login(page);
 
   await page.waitForURL('/dashboard');
